chore: rename wallet connect import context directory (#858)
